+++ src/main/java/com/simonconrad/fireballpredictor/math/BlockStateSnapshot.java
@@ -1,13 +1,13 @@
 package com.simonconrad.fireballpredictor.math;
 
-import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
-import net.minecraft.block.entity.BlockEntity;
-import net.minecraft.fluid.FluidState;
-import net.minecraft.fluid.Fluids;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.BlockView;
 import net.minecraft.world.World;
+import net.minecraft.block.Blocks;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.BlockState;
+import net.minecraft.fluid.FluidState;
+import net.minecraft.fluid.Fluids;
 import org.jetbrains.annotations.Nullable;
 
 public class BlockStateSnapshot implements BlockView {