+++ src/main/java/com/simonconrad/fireballpredictor/network/FireballPowerPayload.java
@@ -1,8 +1,8 @@
 package com.simonconrad.fireballpredictor.network;
 
 import net.minecraft.network.RegistryByteBuf;
-import net.minecraft.network.codec.PacketCodec;
 import net.minecraft.network.codec.PacketCodecs;
+import net.minecraft.network.codec.PacketCodec;
 import net.minecraft.network.packet.CustomPayload;
 import net.minecraft.util.Identifier;